feat(nutrition-domain): extract pure storage-core into domain package (Phase 7 / PR 2)
Продовжує PR #534. Витягнуто pure-частини зі стораж-шару модуля
Харчування у `@sergeant/nutrition-domain`, щоб `apps/mobile` міг
скласти свій MMKV-адаптер поверх тих самих normalize/mutation-хелперів
без копіпасти з `apps/web`.

Нові pure-модулі у пакеті:
- `nutritionTypes` — типи (Meal, NutritionLog, NutritionDay, Pantry,
  NutritionPrefs, MealTemplate, DaySummary, MealSearchResult, MacrosRow,
  MealMacroSource, MealSource, NutritionGoal) + константи LS-ключів
  (NUTRITION_PANTRIES_KEY, NUTRITION_ACTIVE_PANTRY_KEY,
  NUTRITION_PREFS_KEY, NUTRITION_LOG_KEY)
- `nutritionLog` — normalizeMeal / normalizeNutritionLog /
  addLogEntry / removeLogEntry / updateLogEntry / getDayMacros /
  getDaySummary / addDaysISODate / duplicatePreviousDayMeals /
  mergeNutritionLogs / searchMealsByName / getMacrosForDateRange /
  estimateLogBytes / trimLogOldestDays
- `nutritionPantries` — makeDefaultPantry / normalizePantries /
  updatePantry
- `nutritionPrefs` — defaultNutritionPrefs / normalizeNutritionPrefs
  (витягнуто pure-частину з старого loadNutritionPrefs)
- `waterLog` — WATER_LOG_KEY / normalizeWaterLog / getTodayWaterMl /
  addWaterMl / resetTodayWater / WaterLog
- `shoppingList` — SHOPPING_LIST_KEY / normalizeShoppingList /
  toggleShoppingItem / removeCheckedItems / getCheckedItems /
  getTotalCount + types

Web I/O-адаптери (`apps/web/src/modules/nutrition/lib/*Storage.ts`)
тепер тонкі — лишається лише `load*`/`persist*`/`save*` поверх
`createModuleStorage`, а всі типи і чиста логіка реекспортуються з
`@sergeant/nutrition-domain`. Публічна поверхня цих файлів не
змінилась — існуючі імпорти `../lib/nutritionStorage.js`,
`../lib/waterStorage.js`, `../lib/shoppingListStorage.js` продовжують
працювати без змін.

Перевірено: `pnpm turbo run test typecheck lint` для
`@sergeant/{shared,nutrition-domain,web}` — 9/9 green,
617 web-тестів + 20 тестів пакета.
