+++ packages/nutrition-domain/src/index.ts
@@ -2,15 +2,23 @@
 // бізнес-логіка Харчування, яку споживають `apps/web` і `apps/mobile`
 // без платформних залежностей (`localStorage`, `window`, `document`).
 //
-// Phase 7 / PR 1: перший зріз pure-логіки модуля Харчування —
-// meal-types, pantry-text-parser, merge-items, recipe-ids, food-categories,
-// nutrition-format. Решта шарів (storage, cloud-backup, IndexedDB-recipe-book,
-// stats/day-summary який зараз тісно зв'язаний з localStorage) приходить
-// наступними PR-ами після того, як витягнемо platform-IO за `KVStore`-
-// абстракцію.
+// Phase 7 / PR 1: meal-types, pantry-text-parser, merge-items, recipe-ids,
+// food-categories, nutrition-format.
+// Phase 7 / PR 2: pure core storage shapes + mutations — nutrition-log,
+// water-log, shopping-list, pantries, prefs. `load*`/`persist*` I/O-шари
+// живуть у `apps/web` (createModuleStorage) і пізніше в `apps/mobile`
+// (MMKV) — обидва кладуть ті самі normalize/mutation-функції поверх
+// власного KVStore.
 export * from "./mealTypes.js";
 export * from "./pantryTextParser.js";
 export * from "./mergeItems.js";
 export * from "./recipeIds.js";
 export * from "./foodCategories.js";
 export * from "./nutritionFormat.js";
+
+export * from "./nutritionTypes.js";
+export * from "./nutritionPrefs.js";
+export * from "./nutritionPantries.js";
+export * from "./nutritionLog.js";
+export * from "./waterLog.js";
+export * from "./shoppingList.js";
